# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a MVP for Questions and Answers for https://www.nuorder.com/ page.\nInstall MongoDB to test this service or use Docker executing Compose up in docker-compose.yml in the project folder check the CLI for reference in https://docs.docker.com/compose/reference/up/\nIf executing this service by VS Code, just click on right button on docker-compose.yml and Compose Up",
        "title": "FAQ RESTful Service API",
        "termsOfService": "https://www.nuorder.com/",
        "contact": {
            "name": "NuOrder API Support",
            "email": "flavio.costa@ecore.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.nuorder.com/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/answer": {
            "post": {
                "description": "Create a new Answer with the input paylod\nThe root is the question ID\nThe Parent field is the ID of another question that will be answered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "Create a new Answer item",
                "parameters": [
                    {
                        "description": "Create",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Answer"
                        }
                    }
                }
            }
        },
        "/api/v1/answer/{idquestion}": {
            "get": {
                "description": "Get a list of all answers from a question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "Get all answers from a specific Question",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Answer"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/answer/{id}": {
            "get": {
                "description": "Get an answer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "Get one answer item from the API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Answer"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a new Question and Answer with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "Update a new Question and Answer item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Answer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ObjectIdHex(id), was successful updated!"
                    }
                }
            },
            "delete": {
                "description": "Delete a question and answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "Delete one question and answer item from the API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result: success"
                    }
                }
            }
        },
        "/api/v1/question": {
            "get": {
                "description": "Get a list of all questions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Get a list to all questions from the API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Question"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Question and Answer with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Create a new Question and Answer item",
                "parameters": [
                    {
                        "description": "Create",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                }
            }
        },
        "/api/v1/question/{id}": {
            "get": {
                "description": "Get a question and answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Get one question and answer item from the API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a new Question and Answer with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Update a new Question and Answer item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ObjectIdHex(id), was successful updated!"
                    }
                }
            },
            "delete": {
                "description": "Delete a question and all the answers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Delete one question and  all answers items from the API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result: success"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Answer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "7f484f697ee1283a0ca5a028"
                },
                "likes": {
                    "type": "integer",
                    "example": 10
                },
                "parent": {
                    "type": "string",
                    "example": "5f484f697ee3881a0ca9a037"
                },
                "root": {
                    "type": "string",
                    "example": "5f484f697ee3881a0ca9a037"
                },
                "text": {
                    "type": "string",
                    "example": "How can I remove an item?"
                }
            }
        },
        "models.AnswerRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "7f484f697ee1283a0ca5a028"
                },
                "likes": {
                    "type": "integer",
                    "example": 10
                },
                "parent": {
                    "type": "string",
                    "example": "5f484f697ee3881a0ca9a037"
                },
                "root": {
                    "type": "string",
                    "example": "5f484f697ee3881a0ca9a037"
                },
                "text": {
                    "type": "string",
                    "example": "How can I remove an item?"
                }
            }
        },
        "models.Question": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "5f484f697ee3881a0ca9a037"
                },
                "image": {
                    "type": "string",
                    "example": "http://nuorder.com/images/image.jpg"
                },
                "likes": {
                    "type": "integer",
                    "example": 30
                },
                "text": {
                    "type": "string",
                    "example": "How can I remove an item?"
                }
            }
        }
    }
}